# Optimize Image Stuff By Evaluating Plugs in Global Context