--- a/src/Gaffer/PerformanceMonitor.cpp
+++ b/src/Gaffer/PerformanceMonitor.cpp
@@ -108,6 +108,13 @@ const PerformanceMonitor::Statistics &PerformanceMonitor::plugStatistics( const 
 	return it->second;
 }
 
+const PerformanceMonitor::Statistics &PerformanceMonitor::combinedStatistics() const
+{
+	collate();
+	return m_combinedStatistics;
+}
+
+
 void PerformanceMonitor::processStarted( const Process *process )
 {
 	const IECore::InternedString type = process->type();
@@ -162,6 +169,7 @@ void PerformanceMonitor::collate() const
 		for( StatisticsMap::const_iterator mIt = m.begin(), meIt = m.end(); mIt != meIt; ++mIt )
 		{
 			m_statistics[mIt->first] += mIt->second;
+			m_combinedStatistics += mIt->second;
 		}
 		m.clear();
 	}