--- a/src/GafferImage/ImageNode.cpp
+++ b/src/GafferImage/ImageNode.cpp
@@ -91,7 +91,12 @@ bool ImageNode::enabled() const
 void ImageNode::hash( const Gaffer::ValuePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const
 {
 	const ImagePlug *imagePlug = output->parent<ImagePlug>();
-	if( imagePlug && enabled() )
+	bool enabledValue;
+	{
+		ImagePlug::GlobalScope c( context );
+		enabledValue = enabled();
+	}
+	if( imagePlug && enabledValue )
 	{
 		// We don't call ComputeNode::hash() immediately here, because for subclasses which
 		// want to pass through a specific hash in the hash*() methods it's a waste of time (the
@@ -200,7 +205,13 @@ void ImageNode::compute( ValuePlug *output, const Context *context ) const
 
 	// we're computing part of an ImagePlug
 
-	if( !enabled() )
+	bool enabledValue;
+	{
+		ImagePlug::GlobalScope c( context );
+		enabledValue = enabled();
+	}
+
+	if( !enabledValue )
 	{
 		// disabled nodes just output a default black image.
 		output->setToDefault();