--- a/src/GafferImage/ImageSwitch.cpp
+++ b/src/GafferImage/ImageSwitch.cpp
@@ -41,31 +41,6 @@
 
 using namespace GafferImage;
 
-namespace
-{
-
-/// \todo If we introduce a TemporaryContext base class in Context.h
-/// then this should derive from that.
-struct ImageSwitchIndexContext
-{
-
-	ImageSwitchIndexContext( const Gaffer::Context *context )
-		:	m_context( new Gaffer::Context( *context, Gaffer::Context::Borrowed ) ),
-			m_scopedContext( m_context.get() )
-	{
-		m_context->remove( ImagePlug::channelNameContextName );
-		m_context->remove( ImagePlug::tileOriginContextName );
-	}
-
-	private :
-
-		Gaffer::ContextPtr m_context;
-		Gaffer::Context::Scope m_scopedContext;
-
-};
-
-} // namespace
-
 namespace Gaffer
 {
 
@@ -75,7 +50,7 @@ template<>
 struct SwitchTraits<GafferImage::ImageProcessor>
 {
 
-	typedef ImageSwitchIndexContext IndexContext;
+	typedef ImagePlug::GlobalScope IndexContext;
 
 };