--- a/src/GafferImage/ImageTransform.cpp
+++ b/src/GafferImage/ImageTransform.cpp
@@ -276,7 +276,10 @@ void ImageTransform::hashChannelData( const GafferImage::ImagePlug *parent, cons
 	else
 	{
 		// Rotation of either the input or the resampled input.
-		ImageTransform::hashDataWindow( parent, context, h );
+		{
+			ImagePlug::GlobalScope c( context );
+			ImageTransform::hashDataWindow( parent, context, h );
+		}
 
 		const ImagePlug *samplerImage; M33f samplerMatrix;
 		const Box2i samplerRegion = sampler( op, matrix, resampleMatrix, context->get<V2i>( ImagePlug::tileOriginContextName ), samplerImage, samplerMatrix );