--- a/src/GafferImage/Premultiply.cpp
+++ b/src/GafferImage/Premultiply.cpp
@@ -105,7 +105,12 @@ void Premultiply::processChannelData( const Gaffer::Context *context, const Imag
 		return;
 	}
 
-	ConstStringVectorDataPtr inChannelNamesPtr = inPlug()->channelNamesPlug()->getValue();
+	ConstStringVectorDataPtr inChannelNamesPtr;
+	{
+		ImagePlug::GlobalScope c( context );
+		inChannelNamesPtr = inPlug()->channelNamesPlug()->getValue();
+	}
+
 	const std::vector<std::string> &inChannelNames = inChannelNamesPtr->readable();
 	if ( std::find( inChannelNames.begin(), inChannelNames.end(), alphaChannel ) == inChannelNames.end() )
 	{