--- a/src/GafferImage/Resize.cpp
+++ b/src/GafferImage/Resize.cpp
@@ -258,6 +258,7 @@ IECore::ConstFloatVectorDataPtr Resize::computeChannelData( const std::string &c
 
 const ImagePlug *Resize::source() const
 {
+	ImagePlug::GlobalScope c( Context::current() );
 	if( formatPlug()->getValue().getDisplayWindow() == inPlug()->formatPlug()->getValue().getDisplayWindow() )
 	{
 		return inPlug();