--- a/src/GafferImage/Sampler.cpp
+++ b/src/GafferImage/Sampler.cpp
@@ -46,7 +46,10 @@ Sampler::Sampler( const GafferImage::ImagePlug *plug, const std::string &channel
 	m_channelName( channelName ),
 	m_boundingMode( boundingMode )
 {
-	m_dataWindow = m_plug->dataWindowPlug()->getValue();
+	{
+		ImagePlug::GlobalScope c( Context::current() );
+		m_dataWindow = m_plug->dataWindowPlug()->getValue();
+	}
 
 	// We only store the sample window to be able to perform
 	// validation of the calls made to sample() in debug builds.