… src/GafferImage/Mix.cpp
… src/GafferImage/Offset.cpp
… src/GafferImage/Resample.cpp
… src/GafferImage/Shuffle.cpp
… src/GafferImage/Text.cpp
… src/GafferImage/Unpremultiply.cpp
… src/GafferImage/VectorWarp.cpp
… src/GafferImage/Warp.cpp
… src/GafferOSL/OSLImage.cpp