This is just being put up for comment - my guess is that some of these commits may be something that you think are a good idea anyway, some of them you probably don't want, and some of them you might only want if we can come up with good tests to prove performance.

I'm still testing to what extent I actually need these changes, so you could just ignore this until I've got some better data.

It's really hard to evaluate performance of this stuff in isolation. I was seeing significant costs to evaluating channelNames and dataWindow in a test, but I never saw much cost for format - but this depends entirely on the upstream network - it's tough to say whether you possibly build an upstream network that would make it costly to evaluate format plugs.

For a crazy test that I was benchmarking on, this originally sped up hashing by a factor of 2, and overall runtime by up to 20% - but after doing this work, I took a look at what was left, and cleaned up some of the stuff I was doing ( avoiding OSLImage seems to have been the big one ), and now this stuff seems to matter a lot less.  Which makes some sense, this stuff is only expensive if you have something upstream that blows up the hashing time, but now I'm not sure how useful this is.

I will evaluate further once I've got a full setup together of my current image processing thing.