@johnhaddon 2017-03-28
Thanks Daniel - looks pretty good at first glance. I'll take a closer look when you have your latest results available and we have a better idea of what works and what doesn't. As you say though, it's very dependent on the input graph, and it might be that we're slightly hurting the performance for the simple cases while improving the performance for the complex cases.

One thing I might take a look at is making a new `TemporaryContext` class as a base class for your global context. This would just hide away the `new Context( Borrowed ); Context::Scope` pair that we're repeating everywhere, and give us the possibility to further improve performance later by avoiding the heap allocation for small contexts. I've been meaning to do something like this for a while, so it seems to make sense now if we're going to expose the GlobalContext publicly.