@danieldresser 2017-04-05
Given the dependence on the input graph, I'm a bit unsure how much profiling you want me to go into here.  On my current heavy network in practice, running on the farm, batched to 5 frames at a time to benefit from caching of things that are TimeWarped, I'm seeing:
Before:  Avg: 350 seconds, Max: 453 seconds ( per 5 frames )
After:  Avg: 204 seconds, Max: 281 seconds ( per 5 frames )

I wasn't controlling for farm heterogenity, so the percentages won't be completely accurate, but this approximately matches what I'm seeing locally.  There is also quite a perceptible difference in feel interactively when clicking between expensive to hash things that you've already computed.

I'm not sure what the best way to quantify your fears about downsides in simple cases is - hopefully in simple cases, we're not seeing huge amounts of hashing time?

I wonder if it would be worth you taking a look through these PRs, and sort them into things you think are likely a good idea ( after your TemporaryContext addition ), and things that feel unnecessary to you?  Some of these seem like very probably a generally positive thing to me.  If there are some that feel unnecessary to you, I could try running my big production test without them, and if they turn out to be useful in the case I need I could target those areas a bit more specifically in trying to construct a performance test that isn't quite as complicated?