@danieldresser 2017-04-05
Oh, also, I wonder if I should break out the one commit to performanceMonitor into a separate PR - it just adds an output for total hashing time, which is pretty useful when looking at this stuff.