@danieldresser 2017-04-06
I thought about this for a while, and felt a bit uncertain about the best way to clean it up, and talked to Don a bit about it, and then got distracted and did my last two main optimization goals from earlier:
* driverChannel plug on Median ( PR coming as soon as I figure out a good way to write the unit test.  I've got a nifty little Gaffer network that tests it well, but it would be a bit complicated to explain in code )
* optimizations to Warp ( speculative PR up )

The conclusions from my conversation with Don ( well, I shouldn't blame him for this, my conclusions, with some input from him ):
* GlobalContext is a hilariously vague name.  Also it doesn't indicate that it acts as a Scope.  How about I propose something completely different like WholeImageScope, and see if that suggests anything reasonable to you.
* if the accessors are actually things that you should only use if you're not already in a global context, then the names of them should probably reflect that - dataWindowHashInImageScope or something ugly like that?
* if the calling code needs to be aware of the details of contexts anyways, then it kinda feels like these accessors shouldn't exist.  The only real reason for them is to save typing all over the place?

Don suggested an approach like a generic wrapper function:
```
template<typename PlugType>
typename PlugType::ValueType imageScopeValue(const PlugType& plug)
{
   const Gaffer::Context *c = Gaffer::Context::current();
   Gaffer::Context n (*c, Gaffer::Context::Copied );

   Gaffer::Context::Scope s(&n);

   n.remove("a");
   n.remove("b");

   return plug.getValue();
}
```
called like:
```
imageScopeValue( *dataWindowPlug() );
```

I also thought about a templated wrapper version of getValue that I guess would live right down in Gaffer::Plug, and would instantiate a templated class before calling the standard getValue, so you could just do:
```
dataWindowPlug()->getValue< GlobalContext >();
```

I don't really want to make my life getting this merged more difficult, but it also feels a bit like I'm introducing some weird ugliness that we're going to regret.  If you're happy with it though, I'll go ahead and clean it up as is ( though I do feel like I should put clearer names on those accessors ).

Oh, and one other random thought: if creating contexts is expensive, should GlobalContext check if either of the variables it's affecting exist already, and not create a new context if they don't?  I guess it really should be the client's responsibility not to create a GlobalContext when already in a GlobalContext.